+++ test/integration/check/test_check.py
@@ -5,10 +5,10 @@ import shutil
 from utils import run_command_and_get_output
 
 
-def test_check_too_many_args(tt_cmd, tmpdir):
+def test_check_too_many_args(tt_cmd, tmpdir_with_cfg):
     # Testing with more than one specified files.
     cmd = [tt_cmd, "check", "file1", "file2"]
-    rc, output = run_command_and_get_output(cmd, cwd=tmpdir)
+    rc, output = run_command_and_get_output(cmd, cwd=tmpdir_with_cfg)
     assert rc == 1
     assert re.search(r"currently, you can specify only one instance at a time", output)